Move permission checks to Cas1UserAccessService
Permissions are only used by CAS1

Remove unused CAS1 code from UserAccessService

These functions are never called by endpoints used by CAS1
